move calcFrame10 to score.js + apply CR comments
