+++ src/client/score.js
@@ -41,3 +41,25 @@ function calculScoreTotal(frameTable){
 }
 
 module.exports = calculScoreTotal;
+
+function calcFrame10(frame){
+    var result = 0;
+    result = frame.getC1();
+
+    if(result === 10){ // we got a strike on the first throw
+        result += frame.getC2() + frame.getC3();
+        return result;
+    }
+    
+    result += frame.getC2();
+
+    if (result === 10){ // we got a spare 
+        result += frame.getC3();
+        return result;
+    }
+    else{ // we got a normal round with only 2 throws
+        return result;
+    }
+}
+
+module.exports = calcFrame10;
